+++ src/xcore/xcore.h
@@ -496,6 +496,7 @@ struct xConfig {
 		int chans;
 		int latency;		// ms of sound kept in the ring buffer (see sound.h)
 		int latauto;		// let the emulator find that number by itself
+		unsigned filter:1;	// filter the sound down to the output rate, else just average it
 		sndVolume vol;
 		FILE* wavfile;
 	} snd;